requirements.txt
google-api-python-client==2.147.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.0.0
requests==2.32.3
tinytuya==1.17.4
flask==3.0.0
